+++ src/lang-xpath/main/uk/co/reecedunn/intellij/plugin/xpath/intellij/codeInsight/hints/XPathInlayParameterHintsProvider.kt
@@ -20,6 +20,7 @@ import com.intellij.codeInsight.hints.InlayInfo
 import com.intellij.codeInsight.hints.InlayParameterHintsProvider
 import com.intellij.psi.PsiElement
 import uk.co.reecedunn.intellij.plugin.xdm.functions.op.op_qname_presentation
+import uk.co.reecedunn.intellij.plugin.xdm.types.XdmAttributeNode
 import uk.co.reecedunn.intellij.plugin.xdm.types.XdmElementNode
 import uk.co.reecedunn.intellij.plugin.xdm.types.XsQNameValue
 import uk.co.reecedunn.intellij.plugin.xdm.variables.XdmVariableName
@@ -69,6 +70,7 @@ class XPathInlayParameterHintsProvider : InlayParameterHintsProvider {
                 }
                 else -> null
             }
+            is XdmAttributeNode -> element.nodeName
             is XdmElementNode -> element.nodeName
             else -> null
         }